Forward GIF requests directly to S3
In the past we've detected Gifs in the Lambda function, and then passed them through - however this often hits the 6MB response limit from Lambda. GIFs are often large files, and until we are _actually_ resizing them, we should just forward the requests to S3. This means there's no limit in their size which is a lot better.