Fix proxy file
Seems an update to the AWS SDK has meant that the `proxy-file.js` no longer worked properly due to `s3.makeRequest` changing their behaviour. Odd handling of `this` meant that requests for gifs were returning the error message `this.makeRequest is not a function`.

In addition the proxy-file fallback was never supported by the local-server.js or regular server.js files so this functionality was broken on Docker builds