Add TravisCI testing
This will run the test-filesizes in travis on each build. It also compares the baseline file sizes in `fixtures.json`, so if we ever introduce chances that lead to large file sizes (or smaller) we'll get that information in the travis build (files growing larger than the fixture will also fail travis builds).